App/Client/src/app/calculator/calculator.component.ts
import { Component, OnInit } from '@angular/core';
import { FormBuilder, FormGroup } from '@angular/forms';
import { Router } from '@angular/router';

@Component({
  selector: 'app-calculator',
  templateUrl: './calculator.component.html',
  styleUrls: ['./calculator.component.scss']
})
export class CalculatorComponent implements OnInit {

  title = 'Calculator';
  calculatorForm!: FormGroup;

  constructor(private formBuilder: FormBuilder, private router: Router) { }

  ngOnInit(): void {
    this.initializeCalculatorForm();
  }

  private initializeCalculatorForm(): void {
    this.calculatorForm = this.formBuilder.group({
      occupation: [''],
      sumInsured: [''],
      totalMonthlyExpenses: [''],
      state: [''],
      postCode: ['']
    });
  }

  OnPreviousButtonClick() {
    this.router.navigate(['applicant-detail']);
  }
}
